+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## [2.32.0](https://github.com/aws-powertools/powertools-lambda-typescript/compare/v2.31.0...v2.32.0) (2026-03-19)
+
+### Features
+
+- **commons** add Lambda Metadata Service support ([#5109](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5109)) ([46ffc2a](https://github.com/aws-powertools/powertools-lambda-typescript/commit/46ffc2ad062d57678030834c8e1d2755f848cc7b))
+- **commons** add Lambda Metadata Service support ([#5106](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5106)) ([bdfa583](https://github.com/aws-powertools/powertools-lambda-typescript/commit/bdfa5830ec5a315b5f5e687b2cac876bc2e2a748))
+- **event-handler** add type-safe Store API for request and shared state  ([#5081](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5081)) ([03034db](https://github.com/aws-powertools/powertools-lambda-typescript/commit/03034db1b8d6bb4208c657de4e47b298dd72ebfc))
+- **event-handler** add validation support for REST router ([#4736](https://github.com/aws-powertools/powertools-lambda-typescript/issues/4736)) ([3617c8c](https://github.com/aws-powertools/powertools-lambda-typescript/commit/3617c8c1aaed53f9146189ef83ea4ac248250c14))
+
+### Maintenance
+
+- **commons** bump @aws/lambda-invoke-store from 0.2.3 to 0.2.4 ([#5089](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5089)) ([fe519ab](https://github.com/aws-powertools/powertools-lambda-typescript/commit/fe519abf01c152eca16b93e42b48d77ce1d67c10))
+- **logger** bump @aws/lambda-invoke-store from 0.2.3 to 0.2.4 ([#5089](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5089)) ([fe519ab](https://github.com/aws-powertools/powertools-lambda-typescript/commit/fe519abf01c152eca16b93e42b48d77ce1d67c10))
+- **metrics** bump @aws/lambda-invoke-store from 0.2.3 to 0.2.4 ([#5089](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5089)) ([fe519ab](https://github.com/aws-powertools/powertools-lambda-typescript/commit/fe519abf01c152eca16b93e42b48d77ce1d67c10))
+- **batch** bump @aws/lambda-invoke-store from 0.2.3 to 0.2.4 ([#5089](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5089)) ([fe519ab](https://github.com/aws-powertools/powertools-lambda-typescript/commit/fe519abf01c152eca16b93e42b48d77ce1d67c10))
+- **validation** bump ajv from 8.17.1 to 8.18.0  ([#5021](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5021)) ([a7791e1](https://github.com/aws-powertools/powertools-lambda-typescript/commit/a7791e17b13120c6fff81f6fc2f12f44cc1b504a))
+
+### Bug Fixes
+
+- **event-handler** add overloads to route() for typed validation context ([#5052](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5052)) ([dae12e4](https://github.com/aws-powertools/powertools-lambda-typescript/commit/dae12e4c0c84aee1e9d2373c198b7628c26ee32d))
+- **event-handler** default error handler returns a web Response correctly ([#5024](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5024)) ([f8100e8](https://github.com/aws-powertools/powertools-lambda-typescript/commit/f8100e88fdfb46415b357726fcfffdde40116061))
+- **kafka** handle tombstone events without value ([#5017](https://github.com/aws-powertools/powertools-lambda-typescript/issues/5017)) ([79553c9](https://github.com/aws-powertools/powertools-lambda-typescript/commit/79553c96a404df96f57268309c4468fe6c92b470))
+
+
 ## [2.31.0](https://github.com/aws-powertools/powertools-lambda-typescript/compare/v2.30.2...v2.31.0) (2026-02-10)
 
 ### Features